make http-adapter a base class and make a child class callsed jquery-ajax-adapter that implements a jquery version of the http-adapter
